+++ apis/weatherunderground.js
@@ -42,7 +42,7 @@ class WundergroundAPI
 		const queryUri = "https://api.weather.com/v2/pws/observations/current?apiKey=" + this.apiKey + "&stationId=" + this.location + "&format=json&units=" + this.units + '&numericPrecision=decimal';
 		request(encodeURI(queryUri), function (err, response, body)
 		{
-			if (!err)
+			if (!err && body.length > 0)
 			{
 				// Current weather report
 				try
@@ -61,13 +61,18 @@ class WundergroundAPI
 				} catch (e)
 				{
 					that.log.error("Error retrieving weather report and forecast");
+					that.log.error("Response Object: " + body);
 					that.log.error("Error Message: " + e);
 					callback(e);
 				}
 			}
 			else
 			{
-				that.log.error("Error retrieving weather report and forecast");
+				that.log.error("Weather Underground Request failed");
+				that.log.error("Response statusCode: " + response.statusCode + " statusMessage: " + response.statusMessage);
+				if (response.statusCode == 204) {
+					that.log.error("Check to make sure your PWS is not offline. https://www.wunderground.com/member/devices")
+				}
 				that.log.error("Error Message: " + err);
 				callback(err);
 			}